# Better dependency algorithm