Fixes #12. Implements the algorithm I wrote out in #16.

Note: Increases memory pressure to `O(n^2)` where `n = MaxExpectedConcurrentJobs`.

Comes with some absolutely bizarre performance increases:

Before:

| Method                | Threads | ConcurrentJobs | MaxConcurrentJobs | Waves | Mean       | Error    | StdDev   | Allocated |
|---------------------- |-------- |--------------- |------------------ |------ |-----------:|---------:|---------:|----------:|
| **BenchmarkSequential**   | **0**       | **32**             | **32**                | **1024**  |   **353.2 ms** |  **4.14 ms** |  **3.67 ms** |     **600 B** |
| BenchmarkDependancies | 0       | 32             | 32                | 1024  |   124.6 ms |  1.91 ms |  1.79 ms |     600 B |
| **BenchmarkSequential**   | **0**       | **32**             | **2048**              | **1024**  |   **358.1 ms** |  **2.85 ms** |  **2.67 ms** |     **600 B** |
| BenchmarkDependancies | 0       | 32             | 2048              | 1024  |   128.7 ms |  1.17 ms |  1.04 ms |     600 B |
| **BenchmarkSequential**   | **0**       | **128**            | **32**                | **1024**  | **1,430.5 ms** | **21.28 ms** | **19.90 ms** |     **600 B** |
| BenchmarkDependancies | 0       | 128            | 32                | 1024  |   724.9 ms |  2.56 ms |  2.39 ms | 5532392 B |
| **BenchmarkSequential**   | **0**       | **128**            | **2048**              | **1024**  | **1,433.1 ms** | **13.07 ms** | **12.23 ms** |     **600 B** |
| BenchmarkDependancies | 0       | 128            | 2048              | 1024  |   709.9 ms |  2.37 ms |  2.21 ms |     600 B |

After: 

| Method                | Threads | ConcurrentJobs | MaxConcurrentJobs | Waves | Mean        | Error     | StdDev    | Allocated |
|---------------------- |-------- |--------------- |------------------ |------ |------------:|----------:|----------:|----------:|
| **BenchmarkSequential**   | **0**       | **32**             | **32**                | **1024**  |   **353.34 ms** |  **2.671 ms** |  **2.498 ms** |     **600 B** |
| BenchmarkDependancies | 0       | 32             | 32                | 1024  |    39.77 ms |  0.347 ms |  0.307 ms |     600 B |
| **BenchmarkSequential**   | **0**       | **32**             | **2048**              | **1024**  |   **357.67 ms** |  **2.897 ms** |  **2.710 ms** |     **600 B** |
| BenchmarkDependancies | 0       | 32             | 2048              | 1024  |    42.00 ms |  0.381 ms |  0.357 ms |     600 B |
| **BenchmarkSequential**   | **0**       | **128**            | **32**                | **1024**  | **1,427.78 ms** | **10.038 ms** |  **8.898 ms** |     **600 B** |
| BenchmarkDependancies | 0       | 128            | 32                | 1024  |   187.53 ms |  0.461 ms |  0.431 ms | 5524920 B |
| **BenchmarkSequential**   | **0**       | **128**            | **2048**              | **1024**  | **1,437.30 ms** | **14.805 ms** | **13.124 ms** |     **600 B** |
| BenchmarkDependancies | 0       | 128            | 2048              | 1024  |   172.52 ms |  0.513 ms |  0.454 ms |     600 B |


So yeah, 3-4x overhead improvement when it comes to a long dependency chain. Genuinely not sure where this comes from; I had to re-check twice. If you've got any insight let me know. Either way, hey, I'll take it.