@LilithSilver 2023-10-16
After testing with #20, it performs slightly worse as I expected before. So I expect it's actually just significantly more efficient in the specific case of a direct chain of dependencies, and in a real-world situation would be overall slightly less efficient. 

Before:
| Method         | Threads | ConcurrentJobs | Waves | Degree | EdgeChance | Mean      | Error    | StdDev   | Allocated |
|--------------- |-------- |--------------- |------ |------- |----------- |----------:|---------:|---------:|----------:|
| **BenchmarkGraph** | **0**       | **32**             | **512**   | **4**      | **0.05**       |  **44.98 ms** | **0.888 ms** | **1.813 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **32**             | **512**   | **4**      | **0.2**        |  **43.85 ms** | **0.866 ms** | **1.186 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **128**            | **512**   | **4**      | **0.05**       | **167.69 ms** | **2.150 ms** | **2.011 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **128**            | **512**   | **4**      | **0.2**        | **182.95 ms** | **3.007 ms** | **2.812 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **256**            | **512**   | **4**      | **0.05**       | **421.30 ms** | **3.374 ms** | **3.156 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **256**            | **512**   | **4**      | **0.2**        | **425.82 ms** | **2.140 ms** | **2.002 ms** |     **600 B** |

After:

| Method         | Threads | ConcurrentJobs | Waves | Degree | EdgeChance | Mean      | Error    | StdDev   | Allocated |
|--------------- |-------- |--------------- |------ |------- |----------- |----------:|---------:|---------:|----------:|
| **BenchmarkGraph** | **0**       | **32**             | **512**   | **4**      | **0.05**       |  **56.94 ms** | **1.133 ms** | **1.588 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **32**             | **512**   | **4**      | **0.2**        |  **45.05 ms** | **0.817 ms** | **1.034 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **128**            | **512**   | **4**      | **0.05**       | **212.96 ms** | **3.581 ms** | **3.350 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **128**            | **512**   | **4**      | **0.2**        | **217.80 ms** | **3.953 ms** | **3.697 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **256**            | **512**   | **4**      | **0.05**       | **461.14 ms** | **4.659 ms** | **4.358 ms** |     **600 B** |
| **BenchmarkGraph** | **0**       | **256**            | **512**   | **4**      | **0.2**        | **480.77 ms** | **7.254 ms** | **6.785 ms** |     **600 B** |